requirements.txt
streamlit==1.32.0
opencv-python==4.8.1
face-recognition==1.3.0
fer==22.5.1
numpy==1.26.4
pandas==2.2.0
scikit-learn==1.4.0
google-api-python-client==2.120.0
google-generativeai==0.3.2
Pillow==10.2.0
requests==2.31.0
sqlite3==3.45.0
